Change User model and authentication
The user model was extended. Authentication was changed to use email instead of username. Custom backend and sign up for the user was added.

Users can create their own account by entering their email. The sign-up and login forms were added, as well as the templates for the login and the creation of the account. The default redirections and the location of the templates for login and logout actions are added in settings.py.